# modify freelist_hmap/hashmapGetFreePageIDs with better performance