@nicktming 2020-04-28
```
func main() {
	N := int32(10000)
	fm := make(map[pgid]uint64)
	i := int32(0)
	val := int32(0)
	for i = 0; i < N;  {
		val = rand.Int31n(1000)
		fm[pgid(i)] = uint64(val)
		i += val
	}

	f := freelist{
		forwardMap: fm,
	}
	start := time.Now()
	f.hashmapGetFreePageIDs()
	end := time.Now()
	fmt.Printf("origin time:%v\n", end.Sub(start))

	start = time.Now()
	res := f.newHashmapGetFreePageIDs()
	end = time.Now()
	fmt.Printf("new time:%v\n", end.Sub(start))


	if !sort.SliceIsSorted(res, func(i, j int) bool { return res[i] < res[j] }) {
		panic("pgids not sorted")
	}
}
```

based on the above testing program, when changing N, got the following result compared with two functions

| N     | time_used_origin | time_used_new |
| ---      | ---       | ---      | 
| 100 |    71.608µs    |  9.022µs |
| 1000 |    226.745µs    |  19.359µs |
| 10000 |    1.290836ms    |  77.194µs |
| 100000 |    15.973611ms    |  777.995µs |
| 1000000 |    160.912048ms    |  5.714576ms |
| 10000000 |    1.789962277s    |  99.443407ms |